% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matches.R
\name{find_matches}
\alias{find_matches}
\title{Find Matches in Joined Data}
\usage{
find_matches(joined_data, case_list = rownames(joined_data[["case"]]))
}
\arguments{
\item{joined_data}{The output of the \code{range_join} function}

\item{case_list}{A vector of case IDs to find matches for. If not provided,
it defaults to all cases in the joined data.}
}
\description{
Prints the matched controls for each case in the joined data.
}
\examples{
set.seed(3)
dt <- split_data(demo_data)
rj <- range_join(dt, max_ctrl_per_case = 2L, join_var_range = c(-3, 3))
find_matches(rj, case_list = c(2, 16, 90))
}
\author{
Waldir Leoncio
}
