% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rangejoin_rangejoin}
\alias{print.rangejoin_rangejoin}
\title{Print method for rangejoin_rangejoin}
\usage{
\method{print}{rangejoin_rangejoin}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rangejoin_rangejoin}.}

\item{...}{Additional arguments (not used).}
}
\description{
Prints the structure of the joined data.
}
