% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rangejoin_splitdata}
\alias{print.rangejoin_splitdata}
\title{Print method for rangejoin_splitdata}
\usage{
\method{print}{rangejoin_splitdata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rangejoin_splitdata}.}

\item{...}{Additional arguments (not used).}
}
\description{
Prints the structure of the original data, case data, and
control data.
}
