% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangejoin-package.R
\docType{package}
\name{rangejoin-package}
\alias{rangejoin}
\alias{rangejoin-package}
\title{rangejoin: Matches Controls to Treaments Based on a Range of Values}
\description{
Divides a dataset into control and treatment groups and matches controls to treatments without replacement based on a range of values.
}
\author{
\strong{Maintainer}: Waldir Leoncio \email{w.l.netto@medisin.uio.no}

Other contributors:
\itemize{
  \item Ragnhild Sørum Falk [reviewer]
}

}
\keyword{internal}
