% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split dataset into case and control}
\usage{
split_data(data, split_variable)
}
\arguments{
\item{data}{Dataset}

\item{split_variable}{Binary variable to split the dataset}
}
\value{
A list with two data frames: case and ctrl
}
\description{
Splits a data frame into case and control groups based on a
binary variable. If the variable is not provided, it randomly assigns
"Control" and "Case" labels to the data on a 2:1 ratio.
}
\examples{
split_data(demo_data)
}
