% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rangejoin_rangejoin}
\alias{summary.rangejoin_rangejoin}
\title{Summary Method for Range Join Objects}
\usage{
\method{summary}{rangejoin_rangejoin}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rangejoin_rangejoin} created by the
\code{rangejoin} function.}

\item{...}{Additional arguments (not used).}
}
\description{
Summarize a range join object to show relevant statistics.
}
\examples{
set.seed(3)
dt <- split_data(demo_data)
rj <- range_join(dt, max_ctrl_per_case = 2L, age_range = c(-3, 3))
summary(rj)
}
\author{
Waldir Leoncio
}
