% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_join.R
\name{range_join}
\alias{range_join}
\title{Range join}
\usage{
range_join(data, max_ctrl_per_case, age_range)
}
\arguments{
\item{data}{A list with two data frames: case and ctrl}

\item{max_ctrl_per_case}{Maximum number of controls per case.}

\item{age_range}{A vector with two elements: the minimum and maximum age
difference between cases and controls.}
}
\value{
A list with two data frames: control_data and case.
}
\description{
Performs a range join between case and control data.
}
