% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rangejoin_splitdata}
\alias{summary.rangejoin_splitdata}
\title{Summary Method for Split Data Objects}
\usage{
\method{summary}{rangejoin_splitdata}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rangejoin_splitdata} created by the
\code{split_data} function.}

\item{...}{Additional arguments (not used).}
}
\description{
Summarize a split data object to show case, control, and all data summaries.
}
\examples{
set.seed(3)
dt <- split_data(demo_data)
summary(dt)
}
\author{
Waldir Leoncio
}
